+++ docs/sphinx/install.rst
@@ -201,7 +201,7 @@ To build the CUDA Quantum Python API from source using pip, run the following co
 
     git clone https://github.com/NVIDIA/cuda-quantum.git
     cd cuda-quantum && bash scripts/install_prerequisites.sh
-    pip install .
+    pip install . --user
 
 For more information about building the entire C++ and Python API from source, we refer to the `CUDA Quantum GitHub repository`_.
