+++ python/cudaq/kernel/kernel_builder.py
@@ -14,12 +14,12 @@ from .utils import mlirTypeFromPyType, nvqppPrefix
 from .common.givens import givens_builder
 from .common.fermionic_swap import fermionic_swap_builder
 
-from mlir_cudaq.ir import *
-from mlir_cudaq.passmanager import *
-from mlir_cudaq.execution_engine import *
-from mlir_cudaq.dialects import quake, cc
-from mlir_cudaq.dialects import builtin, func, arith
-from mlir_cudaq._mlir_libs._quakeDialects import cudaq_runtime
+from ..mlir.ir import *
+from ..mlir.passmanager import *
+from ..mlir.execution_engine import *
+from ..mlir.dialects import quake, cc
+from ..mlir.dialects import builtin, func, arith
+from ..mlir._mlir_libs._quakeDialects import cudaq_runtime
 
 qvector = cudaq_runtime.qvector
