+++ python/cudaq/kernel/utils.py
@@ -1,14 +1,14 @@
 # ============================================================================ #
-# Copyright (c) 2022 - 2023 NVIDIA Corporation & Affiliates.                   #
+# Copyright (c) 2022 - 2024 NVIDIA Corporation & Affiliates.                   #
 # All rights reserved.                                                         #
 #                                                                              #
 # This source code and the accompanying materials are made available under     #
 # the terms of the Apache License 2.0 which accompanies this distribution.     #
 # ============================================================================ #
-from mlir_cudaq._mlir_libs._quakeDialects import cudaq_runtime
-from mlir_cudaq.dialects import quake, cc
-from mlir_cudaq.ir import *
-from mlir_cudaq.passmanager import *
+from ..mlir._mlir_libs._quakeDialects import cudaq_runtime
+from ..mlir.dialects import quake, cc
+from ..mlir.ir import *
+from ..mlir.passmanager import *
 import numpy as np
 from typing import Callable
 import ast
